requirements.txt
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.0
et_xmlfile==2.0.0
Faker==30.8.2
folium==0.18.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
xyzservices==2024.9.0
